Use zsys_shutdown() before termination.
The Windows platform was experiencing a WSAStartup assertion in the
fncs_broker. FNCS clients would experience a hang during termination
caused by a poor interaction of the zeromq context being created inside
the libfncs DLL instead of in the application main(). The solution was
to call zsys_shutdown() after zsock_destroy() to solve the hang and
avoid the assert within the broker shutdown.