Add message receipt logging for debugging
Prior to this patch, it has not been possible to have any idea why a co-sim hangs. It just does and the only clues are often in the log files. In the case of the mini_federate problem we’re seeing, turning on logging causes the problem to go away.

To enable some degree of logging, libfncs has been changed to place a timeout on the zmq_poll. If the poll times out, the returned value in time_request is set to zero. This is used by the mini_federate to indicate that the co-sim is stuck.

When the co-sim hangs, the mini-federate examines a receipt vector it has been using to indicate when each expected message has been received. If the co-sim hangs, it will be able to indicate in the log file which messages were received and which are missing. This could be used generally in all federates but works best with the mini_federate since the data exchange is so structured and regimented; it is easy to know which messages should be sent and received by which federates at each time step.