Fix bug prevent correct root message receipt logging
Originally, the logging of the message receipts was dependent on whether the time granted matched the time requested. Since root always requests the end-of-co-sim time, it was never doing this accounting.

This has been corrected by moving this message receipt logging outside that condition which results in the print-out of the number of messages received at every time grant.