Remove limit on high-water mark and linger time
During work in CCSI where the limits of the FNCS messaging were being stress-tested with a minimal federate (the mini_federate), a few ZMQ settings needed to be adjusted to ensure proper broker operation during heavy messaging periods. As the comments in the code indicate, the send and receive buffer size were set to infinite (ZMQ "high-water mark") and the linger time for messages "in-flight".

Adjusting both of these settings eliminated problems of the federation hanging where it appeared all federates were hanging around waiting for something.

These adjustments have the potential to cause problems with resources and performance but until a specific problems is identified, it is difficult to determine what appropriate settings should be.