Add check for extra federates trying to join
Previously, there was no check for federates beyond the count specified attempting to join. If such a situation occurred they would be accepted into the federation (even though we specify the number of federates when launching the co-sim). A check has been added so that extra-joiners will not be admitted.